Методы для автора перенесены в компонент Author в папке API
